changed datatypes in address

+++ deploy/materialized_view_address.sql
@@ -16,8 +16,8 @@ create materialized view ggircs_swrs.address as (
            '//Address[not(ancestor::Stack)]'
            passing source_xml
            columns
-                swrs_facility_id numeric(1000,0) path './ancestor::Facility/../../ReportDetails/FacilityId[normalize-space(.)]|./ancestor::Contact/ancestor::ReportData/ReportDetails/FacilityId[normalize-space(.)]',
-                swrs_organisation_id numeric(1000,0) path './ancestor::Organisation/../../ReportDetails/OrganisationId[normalize-space(.)]|./ancestor::ParentOrganisation/ancestor::ReportData/ReportDetails/OrganisationId[normalize-space(.)]',
+                swrs_facility_id integer path './ancestor::Facility/../../ReportDetails/FacilityId[normalize-space(.)]|./ancestor::Contact/ancestor::ReportData/ReportDetails/FacilityId[normalize-space(.)]',
+                swrs_organisation_id integer path './ancestor::Organisation/../../ReportDetails/OrganisationId[normalize-space(.)]|./ancestor::ParentOrganisation/ancestor::ReportData/ReportDetails/OrganisationId[normalize-space(.)]',
                 path_context varchar(1000) path 'name(./ancestor::VerifyTombstone|./ancestor::RegistrationData)',
                 type varchar(1000) path 'name(..)',
                 contact_idx integer path 'string(count(./ancestor::Contact/preceding-sibling::Contact))' not null,
